requirements.txt
asgiref==3.6.0
certifi==2023.5.7
charset-normalizer==3.1.0
Django==3.2
django-cors-headers==4.0.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
jdatetime==4.1.1
Pillow==9.5.0
PyJWT==2.7.0
pymemcache==4.0.0
pytz==2023.3
requests==2.30.0
sqlparse==0.4.4
urllib3==2.0.2
